Reorganize Step 4 (#17)
While the interactivity built in Step 4 to display the selected email is pretty straightforward, 4 concepts get taught in the process. And w/o clear indications of those concepts, the workshop feels very overwhelming. So this diff first clearly identifies those concepts. Then it reorganizes how the functionality is built so that there's something visible in the app for the majority of the Step instead of building out all this functionality w/ a ta-da at the end.

Fixes #12