Swap steps 9 & 10 (#31)
Learning about styling is more important than what we learn when adding the mark read/unread functionality. So it makes since to swap steps 9 & 10 so that styling comes first. The mark read/unread step does introduce a lot of conditional logic especially in the Exercises so it'll be good to get to. But learning how to style is more important.

I tried to minimize the git changes. The state of the code at the end of Step 10 was still they same, so the code really didn't need to move. I basically just renamed the folders and swapped the READMEs. Most of the work was updating the READMEs to reflect that in-progress state of the code.

Fixes #14