Flip unread property in dataset (#35)
The dataset, REST API & `App` functions all dealt with an `unread` property, which is a negative word. So things get real confusing when we had `{unread: false}`, which really could be better expressed as `{read: true}`.

So this flips the dataset, REST API & `App` functions to deal with `read` property instead of `unread` property. Nearly almost every source file had to change.

Fixes #20